finished test for the player ingester, par3model and some of player service
